Web-App/requirements.txt
alembic==1.3.1
Click==7.0
dominate==2.4.0 #diff
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.10.3 ##diff
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.11
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
Flask-Mail==0.9.1
PyJWT==2.3.0
email-validator==1.1.3 #2.3.x
debugpy